+++ packages/react/form/src/index.ts
@@ -3,3 +3,5 @@ export * from './lib/react-form';
 export * from '@paintbox/react-checkbox';
 export * from '@paintbox/react-time-picker';
 export * from '@paintbox/react-search-bar';
+export * from '@paintbox/react-select-option';
+export * from '@paintbox/react-currency-input';