+++ packages/react/time-picker/src/lib/react-time-picker.tsx
@@ -3,17 +3,18 @@
 // @ts-nocheck
 
 // FIXME: 在專案中使用這個套件時，react-datepicker 的 default import 會是整個 module，而不是 react-datepicker 中預設 export 的元件。為了暫時解決這個問題，在 JSX 中是 call ReactDatePicker.default。不過在 storybook 中會報錯，目前僅以文字說明。
+// TODO: 要可以清空
 
 import React from 'react';
 import ReactDatePicker from 'react-datepicker';
 import styled from '@emotion/styled';
 
 import { Clock, Calendar3 } from '@paintbox/react-foundation';
-import { FormFieldBaseProps, ChangeHandlerParams } from './declarations';
+import { FormFieldBaseProps, ChangeHandlerParams } from '@paintbox/react-base';
 import {
   convertUTCToLocalDate,
   convertLocalToUTCDate,
-} from '@paintbox/util-datetime';
+} from '@paintbox/util-i18n';
 
 import 'react-datepicker/dist/react-datepicker.css';
 import './custom-react-datepicker.css';
@@ -108,8 +109,6 @@ export function ReactTimePicker<NameType>({
     </Container>
   );
 }
-
-export default ReactTimePicker;
 
 export const TimepickerSample = () => {
   const [value, setValue] = React.useState<number | null>(null);