+++ packages/util/datetime/src/lib/util-datetime.spec.ts
@@ -1,7 +1,0 @@
-import { utilDatetime } from './util-datetime';
-
-describe('utilDatetime', () => {
-  it('should work', () => {
-    expect(utilDatetime()).toEqual('util-datetime');
-  });
-});