+++ packages/util/i18n/src/lib/util-i18n.spec.ts
@@ -1,0 +1,7 @@
+import { utilI18n } from './util-i18n';
+
+describe('utilI18n', () => {
+  it('should work', () => {
+    expect(utilI18n()).toEqual('util-i18n');
+  });
+});